### 2024-02-23 #6402 @editorialbot
[PRE REVIEW]: piglot: an Open-source Package for Derivative-free Optimisation of Numerical Responses
**Submitting author:** @ruicoelhopedro (<a href="http://orcid.org/0000-0001-9989-964X">Rui Pedro Cardoso Coelho</a>)
**Repository:** https://github.com/CM2S/piglot
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
**Managing EiC:** Daniel Katz
