### 2024-02-23 8a52391
Fix invalid DOI in JOSS paper
