+++ docs/joss_paper/references.bib
@@ -88,7 +88,7 @@ journal = {International Journal of Solids and Structures},
 pages = {112488},
 year = {2023},
 issn = {0020-7683},
-doi = {https://doi.org/10.1016/j.ijsolstr.2023.112488},
+doi = {10.1016/j.ijsolstr.2023.112488},
 url = {https://www.sciencedirect.com/science/article/pii/S0020768323003852},
 author = {A. Francisca {Carvalho Alves} and Bernardo P. Ferreira and F.M. Andrade Pires},
 keywords = {Rubber-toughened amorphous polymers, Polycarbonate, Visco-elastic visco-plastic, Internal particle cavitation, Constitutive model, Finite element analysis, Parameter identification}
