### 2024-02-26 #6402 @ruicoelhopedro
List of potential reviewers:
- [Extraweich](https://reviewers.joss.theoj.org/reviewers/2571)
- [pprachas](https://reviewers.joss.theoj.org/reviewers/2955)
- [jbussemaker](https://reviewers.joss.theoj.org/reviewers/2802)
- [sgbaird](https://reviewers.joss.theoj.org/reviewers/2899)
