### 2024-03-08 3f7e2e8
Fix standardisation for test dataset
