+++ piglot/optimisers/botorch/bayes.py
@@ -1,7 +1,9 @@
 """Main optimiser classes for using BoTorch with piglot"""
-from typing import Tuple, List, Union
+from typing import Tuple, List, Union, Type
 from multiprocessing.pool import ThreadPool as Pool
+import os
 import warnings
+import dataclasses
 import numpy as np
 import torch
 from scipy.stats import qmc
@@ -11,12 +13,17 @@ from botorch.fit import fit_gpytorch_mll
 from botorch.optim import optimize_acqf
 from botorch.models.model import Model
 from botorch.models import SingleTaskGP
+from botorch.models.converter import batched_to_model_list
 from botorch.acquisition.acquisition import AcquisitionFunction
 from botorch.acquisition import UpperConfidenceBound, qUpperConfidenceBound
 from botorch.acquisition import ExpectedImprovement, qExpectedImprovement
 from botorch.acquisition import ProbabilityOfImprovement, qProbabilityOfImprovement
 from botorch.acquisition.objective import GenericMCObjective
 from botorch.acquisition.knowledge_gradient import qKnowledgeGradient
+from botorch.acquisition.multi_objective.objective import GenericMCMultiOutputObjective
+from botorch.acquisition.multi_objective.logei import qLogExpectedHypervolumeImprovement
+from botorch.utils.multi_objective.box_decompositions.non_dominated import \
+    FastNondominatedPartitioning, NondominatedPartitioning
 from botorch.sampling import SobolQMCNormalSampler
 from piglot.objective import Objective, GenericObjective, ObjectiveResult
 from piglot.optimisers.botorch.dataset import BayesDataset
@@ -44,6 +51,27 @@ def fit_mll_pytorch_loop(mll: ExactMarginalLogLikelihood, n_iters: int = 100) ->
         optimizer.step()
     mll.model.eval()
     mll.model.likelihood.eval()
+
+
+@dataclasses.dataclass
+class MultiObjectiveData:
+    """Data class for multi-objective optimisation."""
+    ref_point: torch.Tensor
+    mapping: List[List[int]]
+    partitioning: NondominatedPartitioning
+    partitioning_type: Type[NondominatedPartitioning]
+
+    def __init__(
+        self,
+        mapping: List[List[int]],
+        partitioning_type: Type[NondominatedPartitioning],
+        ref_point: torch.Tensor = None,
+        partitioning: NondominatedPartitioning = None,
+    ) -> None:
+        self.ref_point = ref_point
+        self.mapping = mapping
+        self.partitioning = partitioning
+        self.partitioning_type = partitioning_type
 
 
 class BayesianBoTorch(Optimiser):
@@ -79,6 +107,7 @@ class BayesianBoTorch(Optimiser):
             raise RuntimeError(f"Unkown acquisition function {self.acquisition}")
         if not self.acquisition.startswith('q') and self.q != 1:
             raise RuntimeError("Can only use q != 1 for quasi-Monte Carlo acquisitions")
+        self.mo_data: MultiObjectiveData = None
         torch.set_num_threads(1)
 
     def _validate_problem(self, objective: Objective) -> None:
@@ -96,6 +125,8 @@ class BayesianBoTorch(Optimiser):
         model: Model,
         n_dim: int,
     ) -> Tuple[AcquisitionFunction, int, int]:
+        if self.objective.multi_objective:
+            return self._build_acquisition_mo(dataset, model, n_dim)
         if self.objective.composition:
             return self._build_acquisition_composite(dataset, model, n_dim)
         return self._build_acquisition_scalar(dataset, model, n_dim)
@@ -116,7 +147,7 @@ class BayesianBoTorch(Optimiser):
         except botorch.exceptions.ModelFittingError:
[…]
