+++ piglot/objective.py
@@ -161,8 +161,37 @@ class ObjectiveResult:
     values: List[np.ndarray]
     variances: Optional[List[np.ndarray]] = None
 
+    @staticmethod
+    def __mc_variance(
+        values: np.ndarray,
+        variances: np.ndarray,
+        composition: Composition,
+        num_samples: int = 1000,
+    ) -> Tuple[float, float]:
+        """Compute the objective variance using Monte Carlo (using fixed base samples).
+
+        Parameters
+        ----------
+        values : np.ndarray
+            Objective values.
+        variances : np.ndarray
+            Objective variances.
+        composition : Composition
+            Composition functional to use.
+        num_samples : int, optional
+            Number of Monte Carlo samples, by default 1000.
+
+        Returns
+        -------
+        float
+            Estimated variance of the objective.
+        """
+        biased = [norm.rvs(loc=0, scale=1) for _ in range(num_samples)]
+        mc_objectives = [composition(values + bias * np.sqrt(variances)) for bias in biased]
+        return np.var(mc_objectives)
+
     def scalarise(self, composition: Composition = None) -> float:
-        """Scalarise the result.
+        """Scalarise the result under noise-free single-objective optimisation.
 
         Parameters
         ----------
@@ -178,8 +207,8 @@ class ObjectiveResult:
             return composition(np.concatenate(self.values))
         return np.mean(self.values)
 
-    def scalarise_mo(self, composition: Composition = None) -> float:
-        """Scalarise the result.
+    def scalarise_mo(self, composition: Composition = None) -> List[float]:
+        """Pseudo-scalarise the result under noise-free multi-objective optimisation.
 
         Parameters
         ----------
@@ -188,15 +217,15 @@ class ObjectiveResult:
 
         Returns
         -------
-        float
-            Scalarised result.
+        List[float]
+            Pseudo-scalarised result.
         """
         if composition is not None:
             return [composition(vals) for vals in self.values]
         return [np.mean(vals) for vals in self.values]
 
     def scalarise_stochastic(self, composition: Composition = None) -> Tuple[float, float]:
-        """Scalarise the result.
+        """Scalarise the result under stochastic single-objective optimisation.
 
         Parameters
         ----------
@@ -211,11 +240,28 @@ class ObjectiveResult:
         if composition is not None:
             values = np.concatenate(self.values)
             variances = np.concatenate(self.variances)
-            # Compute the objective variance using Monte Carlo (using fixed base samples)
-            biased = [norm.rvs(loc=0, scale=1) for _ in range(1000)]
-            mc_objectives = [composition(values + bias * np.sqrt(variances)) for bias in biased]
-            return composition(values), np.var(mc_objectives)
+            return composition(values), self.__mc_variance(values, variances, composition)
         return np.mean(self.values), np.sum(self.variances)
+
+    def scalarise_mo_stochastic(self, composition: Composition = None) -> List[Tuple[float, float]]:
+        """Pseudo-scalarise the result under stochastic multi-objective optimisation.
+
+        Parameters
+        ----------
+        composition : Composition, optional
+            Composition functional to use, by default None.
+
+        Returns
+        -------
+        List[Tuple[float, float]]
+            Pseudo-scalarised  mean and variance.
+        """
+        if composition is not None:
+            return [
+                (composition(vals), self.__mc_variance(vals, vars, composition))
+                for vals, vars in zip(self.values, self.variances)
+            ]
+        return [(np.mean(vals), np.sum(vars)) for vals, vars in zip(self.values, self.variances)]
 
 
 class GenericObjective(Objective):
@@ -251,6 +297,8 @@ class GenericObjective(Objective):
                 if self.multi_objective:
                     for i in range(self.num_objectives):
[…]
