### 2024-03-27 80d82fd
Remove Figure 1 from JOSS paper (#20)
