### 2024-03-27 #6402 @ruicoelhopedro
> @ruicoelhopedro - thanks for this submission. I will look for an editor next. While I do so, I suggest you remove figure 1 from the paper, as it doesn't really add to the paper, isn't mentioned in the text, and isn't needed since the logo is also in figure 2. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Done, thank you for the feedback!
