### 2024-03-28 1a955ac
Add tests for composite posteriors
