+++ piglot/optimisers/botorch/bayes.py
@@ -1,26 +1,98 @@
 """Main optimiser classes for using BoTorch with piglot"""
-from typing import Tuple, List, Union
+from typing import Tuple, List, Union, Optional, Any
 from multiprocessing.pool import ThreadPool as Pool
 import warnings
 import numpy as np
 import torch
 from scipy.stats import qmc
+from torch.quasirandom import SobolEngine
 from gpytorch.mlls import ExactMarginalLogLikelihood
+from gpytorch.models import ExactGP
 import botorch
+from botorch.exceptions import UnsupportedError
 from botorch.fit import fit_gpytorch_mll
 from botorch.optim import optimize_acqf
 from botorch.models.model import Model
-from botorch.models import SingleTaskGP
+from botorch.models import SingleTaskGP, MultiTaskGP
 from botorch.acquisition.acquisition import AcquisitionFunction
 from botorch.acquisition import UpperConfidenceBound, qUpperConfidenceBound
 from botorch.acquisition import ExpectedImprovement, qExpectedImprovement
+from botorch.acquisition import LogExpectedImprovement, qLogExpectedImprovement
 from botorch.acquisition import ProbabilityOfImprovement, qProbabilityOfImprovement
 from botorch.acquisition.objective import GenericMCObjective
 from botorch.acquisition.knowledge_gradient import qKnowledgeGradient
-from botorch.sampling import SobolQMCNormalSampler
+from botorch.sampling import SobolQMCNormalSampler, MultivariateNormalQMCEngine
+from botorch.sampling.normal import NormalMCSampler
+from botorch.posteriors import Posterior
 from piglot.objective import Objective, GenericObjective, ObjectiveResult
 from piglot.optimisers.botorch.dataset import BayesDataset
 from piglot.optimiser import Optimiser
+
+
+AVAILABLE_ACQUISITIONS = (
+    'ucb',
+    'ei',
+    'pi',
+    'kg',
+    'qucb',
+    'qei',
+    'qpi',
+    'qkg',
+    'logei',
+    'qlogei',
+)
+
+
+class SobolQMCMultivariateNormalSampler(NormalMCSampler):
+    r"""Sampler for quasi-MC N(0,sigma) base samples using Sobol sequences.
+    """
+
+    def __init__(
+        self,
+        sample_shape: torch.Size,
+        model: ExactGP,
+        seed: Optional[int] = None,
+        **kwargs: Any,
+    ) -> None:
+        super().__init__(sample_shape=sample_shape, seed=seed, **kwargs)
+        self.cov = torch.corrcoef(model.train_targets)
+        if len(self.cov.shape) == 0:
+            self.cov = torch.eye(1)
+
+    def _construct_base_samples(self, posterior: Posterior) -> None:
+        r"""Generate quasi-random Normal base samples (if necessary).
+
+        This function will generate a new set of base samples and set the
+        `base_samples` buffer if one of the following is true:
+
+        - the MCSampler has no `base_samples` attribute.
+        - the output of `_get_collapsed_shape` does not agree with the shape of
+            `self.base_samples`.
+
+        Args:
+            posterior: The Posterior for which to generate base samples.
+        """
+        target_shape = self._get_collapsed_shape(posterior=posterior)
+        if self.base_samples is None or self.base_samples.shape != target_shape:
+            base_collapsed_shape = target_shape[len(self.sample_shape):]
+            output_dim = base_collapsed_shape.numel()
+            if output_dim > SobolEngine.MAXDIM:
+                raise UnsupportedError(
+                    "SobolQMCSampler only supports dimensions "
+                    f"`q * o <= {SobolEngine.MAXDIM}`. Requested: {output_dim}"
+                )
[…]
