### 2024-03-29 ee80d5e
Expand composite observations to full size
