+++ piglot/objectives/compositions.py
@@ -1,4 +1,5 @@
 """Module for compositions of objectives."""
+from typing import List, Tuple
 import numpy as np
 import torch
 from piglot.objective import Composition
@@ -6,21 +7,6 @@ from piglot.objective import Composition
 
 class MSEComposition(Composition):
     """Mean squared error outer composite function with gradients."""
-
-    def composition(self, inner: np.ndarray) -> float:
-        """Compute the MSE outer function of the composition.
-
-        Parameters
-        ----------
-        inner : np.ndarray
-            Return value from the inner function.
-
-        Returns
-        -------
-        float
-            Composition result.
-        """
-        return np.mean(np.square(inner), axis=-1)
 
     def composition_torch(self, inner: torch.Tensor) -> torch.Tensor:
         """Compute the MSE outer function of the composition with gradients.
@@ -41,23 +27,8 @@ class MSEComposition(Composition):
 class MAEComposition(Composition):
     """Mean absolute error outer composite function with gradients."""
 
-    def composition(self, inner: np.ndarray) -> float:
-        """Compute the CAE outer function of the composition.
-
-        Parameters
-        ----------
-        inner : np.ndarray
-            Return value from the inner function.
-
-        Returns
-        -------
-        float
-            Composition result.
-        """
-        return np.mean(np.abs(inner), axis=-1)
-
     def composition_torch(self, inner: torch.Tensor) -> torch.Tensor:
-        """Compute the CAE outer function of the composition with gradients.
+        """Compute the MAE outer function of the composition with gradients.
 
         Parameters
         ----------
@@ -76,3 +47,114 @@ AVAILABLE_COMPOSITIONS = {
     'mse': MSEComposition,
     'mae': MAEComposition,
 }
+
+
+class UnflattenUtility:
+    """Utility for unflattening a set of responses (with gradients)."""
+
+    def __init__(self, lengths: List[int]):
+        self.lengths = lengths
+        self.indices = np.cumsum([0] + lengths)
+
+    def unflatten(self, data: np.ndarray) -> List[np.ndarray]:
+        """Unflatten a vector containing a set of responses.
+
+        Parameters
+        ----------
+        data : np.ndarray
+            Flattened data.
+
+        Returns
+        -------
+        List[np.ndarray]
+            List of unflatten responses.
+        """
+        return [data[..., self.indices[i]:self.indices[i+1]] for i in range(len(self.lengths))]
+
+    def unflatten_torch(self, data: torch.Tensor) -> List[torch.Tensor]:
+        """Unflatten a vector containing a set of responses with gradients.
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Flattened data.
+
+        Returns
+        -------
+        List[torch.Tensor]
+            List of unflatten responses (with gradients).
+        """
+        return [data[..., self.indices[i]:self.indices[i+1]] for i in range(len(self.lengths))]
+
+
+class FixedLengthTransformer:
+    """Utility for transforming a response into a fixed-length format and back."""
+
+    def __init__(self, n_points: int):
+        self.n_points = n_points
+
+    def length(self) -> int:
+        """Return the length of the fixed-length response.
+
+        Returns
+        -------
+        int
+            Length of the fixed-length response.
+        """
+        return self.n_points + 2
+
+    def transform(self, time: np.ndarray, data: np.ndarray) -> np.ndarray:
+        """Transform a response into a fixed-length format.
+
+        Parameters
+        ----------
+        time : np.ndarray
+            Time points to transform.
+        data : np.ndarray
+            Data points to transform.
+
+        Returns
+        -------
+        np.ndarray
[…]
