+++ piglot/objectives/design.py
@@ -5,10 +5,12 @@ from abc import ABC, abstractmethod
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
+import torch
 from piglot.parameter import ParameterSet
 from piglot.solver import read_solver
 from piglot.solver.solver import Solver, OutputResult
-from piglot.objective import DynamicPlotter, GenericObjective, ObjectiveResult
+from piglot.objective import Composition, DynamicPlotter, GenericObjective, ObjectiveResult
+from piglot.objectives.compositions import UnflattenUtility, FixedLengthTransformer
 from piglot.utils.assorted import stats_interp_to_common_grid
 from piglot.utils.solver_utils import load_module_from_file
 
@@ -16,62 +18,107 @@ from piglot.utils.solver_utils import load_module_from_file
 class Quantity(ABC):
     """Abstract class for quantities to be computed from a response."""
 
-    @abstractmethod
-    def compute(self, result: OutputResult) -> float:
+    def compute(self, time: np.ndarray, data: np.ndarray) -> float:
         """Compute this quantity for a given response.
 
         Parameters
         ----------
-        result : OutputResult
-            Output result to compute the quantity for.
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
 
         Returns
         -------
         float
             Quantity value.
+        """
+        return self.compute_torch(torch.from_numpy(time), torch.from_numpy(data)).item()
+
+    @abstractmethod
+    def compute_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Compute this quantity for a given response (with gradients).
+
+        Parameters
+        ----------
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
+
+        Returns
+        -------
+        torch.Tensor
+            Quantity value (with gradients).
         """
 
 
 class MaxQuantity(Quantity):
     """Maximum value of a response."""
 
-    def compute(self, result: OutputResult) -> float:
-        """Get the maximum of a given response.
+    def compute_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Get the maximum of a given response (with gradients).
 
         Parameters
         ----------
-        result : OutputResult
-            Output result to compute the quantity for.
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
 
         Returns
         -------
-        float
-            Quantity value.
+        torch.Tensor
+            Quantity value (with gradients).
         """
-        return np.max(result.get_data())
+        return torch.max(data, dim=-1)[0]
+
+
+class MinQuantity(Quantity):
+    """Minimum value of a response."""
+
+    def compute_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Get the minimum of a given response (with gradients).
+
+        Parameters
+        ----------
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
+
+        Returns
+        -------
+        torch.Tensor
+            Quantity value (with gradients).
+        """
+        return torch.min(data, dim=-1)[0]
 
 
 class IntegralQuantity(Quantity):
     """Integral of a response."""
 
-    def compute(self, result: OutputResult) -> float:
-        """Get the integral of a given response.
+    def compute_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Get the integral of a given response (with gradients).
 
         Parameters
         ----------
-        result : OutputResult
-            Output result to compute the quantity for.
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
 
         Returns
         -------
-        float
-            Quantity value.
[…]
