### 2024-04-02 #6402 @danielskatz
👋 @ppxasjsm - this is just a ping to see if you can edit this, as mentioned [above](https://github.com/openjournals/joss-reviews/issues/6402#issuecomment-2020875589).
