+++ piglot/utils/composition/responses.py
@@ -1,0 +1,395 @@
+"""Module with utilities for transforming responses under compositions."""
+from typing import List, Tuple
+from abc import ABC, abstractmethod
+import numpy as np
+import torch
+from piglot.objective import Composition, ObjectiveResult
+from piglot.solver.solver import OutputResult
+from piglot.utils.reductions import Reduction
+
+
+class FlattenUtility(ABC):
+    """Utility for flattening a response into a fixed-size vector (with gradients)."""
+
+    @abstractmethod
+    def length(self) -> int:
+        """Return the length of the flattened vector.
+
+        Returns
+        -------
+        int
+            Length of the flattened vector.
+        """
+
+    @abstractmethod
+    def flatten_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Flatten a response into a single vector (with gradients).
+
+        Parameters
+        ----------
+        time : torch.Tensor
+            Time points of the response.
+        data : torch.Tensor
+            Data points of the response.
+
+        Returns
+        -------
+        torch.Tensor
+            Flattened responses.
+        """
+
+    @abstractmethod
+    def unflatten_torch(self, data: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        """Unflatten a vector containing a response (with gradients).
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Flattened responses.
+
+        Returns
+        -------
+        Tuple[torch.Tensor, torch.Tensor]
+            List of responses.
+        """
+
+    def flatten(self, time: np.ndarray, data: np.ndarray) -> np.ndarray:
+        """Flatten a response into a single vector.
+
+        Parameters
+        ----------
+        time : np.ndarray
+            Time points of the response.
+        data : np.ndarray
+            Data points of the response.
+
+        Returns
+        -------
+        np.ndarray
+            Flattened responses.
+        """
+        return self.flatten_torch(torch.from_numpy(time), torch.from_numpy(data)).numpy(force=True)
+
+    def unflatten(self, data: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        """Unflatten a vector containing a response.
+
+        Parameters
+        ----------
+        data : np.ndarray
+            Flattened responses.
+
+        Returns
+        -------
+        Tuple[np.ndarray, np.ndarray]
+            List of responses.
+        """
+        time, values = self.unflatten_torch(torch.from_numpy(data))
+        return time.numpy(force=True), values.numpy(force=True)
+
+
+class FixedFlatteningUtility(FlattenUtility):
+    """Response flattening utility for fixed time grids."""
+
+    def __init__(self, time_grid: np.ndarray):
+        self.time_grid = time_grid
+
+    def length(self) -> int:
+        """Return the length of the flattened vector.
+
+        Returns
+        -------
+        int
+            Length of the flattened vector.
+        """
+        return len(self.time_grid)
+
+    def flatten_torch(self, time: torch.Tensor, data: torch.Tensor) -> torch.Tensor:
+        """Flatten a response into a single vector (with gradients).
+
+        Parameters
+        ----------
+        time : torch.Tensor
+            Time points of the response.
+        data : torch.Tensor
+            Data points of the response.
+
+        Returns
+        -------
+        torch.Tensor
+            Flattened responses.
+        """
+        if time.shape[-1] != len(self.time_grid):
+            raise ValueError("Time grid does not match the expected length.")
+        if time.shape != data.shape:
+            raise ValueError("Mismatched time and data shapes are not supported.")
+        return data
+
+    def unflatten_torch(self, data: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        """Unflatten a vector containing a response (with gradients).
+
+        Parameters
[…]
