### 2024-04-10 4e03f5f
Add option to read output fields from a script for Links solvers
