### 2024-04-11 542a3bc
Fix checking Links output data on non-existing files
