### 2024-04-11 2da2f75
Expand the parameters to match the sizes of the time grids in compositions
