### 2024-04-16 59b360e
Fix wrong noisy assignment in BoTorch
