### 2024-04-17 cb9cdb6
Refactor and simplify BoTorch dataset standardisation
