+++ piglot/optimisers/botorch/dataset.py
@@ -13,7 +13,6 @@ class BayesDataset:
             self,
             n_dim: int,
             n_outputs: int,
-            bounds: np.ndarray,
             export: str = None,
             dtype: torch.dtype = torch.float64,
             std_tol: float = 1e-6,
@@ -27,12 +26,27 @@ class BayesDataset:
         self.params = torch.empty((0, n_dim), dtype=dtype, device=device)
         self.values = torch.empty((0, n_outputs), dtype=dtype, device=device)
         self.variances = torch.empty((0, n_outputs), dtype=dtype, device=device)
-        self.lbounds = torch.tensor(bounds[:, 0], dtype=dtype, device=device)
-        self.ubounds = torch.tensor(bounds[:, 1], dtype=dtype, device=device)
+        self.outcome_mask = torch.empty(n_outputs, dtype=torch.bool, device=device)
+        self.outcome_means = torch.empty(n_outputs, dtype=dtype, device=device)
+        self.outcome_stds = torch.empty(n_outputs, dtype=dtype, device=device)
         self.export = export
         self.std_tol = std_tol
         self.def_variance = def_variance
         self.device = device
+
+    def __update_stats(self) -> None:
+        """Update the statistics of the dataset."""
+
+        # Get observation statistics
+        y_avg = torch.mean(self.values, dim=-2)
+        y_std = torch.std(self.values, dim=-2)
+        # Build mask of points with near-null variance
+        y_abs_avg = torch.mean(torch.abs(self.values), dim=-2)
+        mask = torch.abs(y_std / y_abs_avg) > self.std_tol
+        # Update the dataset statistics
+        self.outcome_mask = mask
+        self.outcome_means = y_avg
+        self.outcome_stds = y_std
 
     def load(self, filename: str) -> None:
         """Load data from a given input file.
@@ -48,6 +62,7 @@ class BayesDataset:
         for point in joint:
             point_np = point.numpy()
             self.push(point_np[:idx1], point_np[idx1:idx2], point_np[idx2:])
+        self.__update_stats()
 
     def save(self, filename: str) -> None:
         """Save all dataset data to a file.
@@ -83,161 +98,72 @@ class BayesDataset:
         self.values = torch.cat([self.values, torch_value.unsqueeze(0)], dim=0)
         self.variances = torch.cat([self.variances, torch_variance.unsqueeze(0)], dim=0)
         self.n_points += 1
+        self.__update_stats()
         # Update the dataset file after every push
         if self.export is not None:
             self.save(self.export)
 
-    def get_obervation_stats(self) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """Return statistics of the observations.
-
-        Returns
-        -------
-        Tuple[torch.Tensor, torch.Tensor, torch.Tensor]
-            Mask, average and standard deviation of the observations.
-
-        Raises
-        ------
-        RuntimeError
-            When all observed points are equal.
-        """
-        # Get observation statistics
-        y_avg = torch.mean(self.values, dim=-2)
-        y_std = torch.std(self.values, dim=-2)
-        # Build mask of points with near-null variance
-        y_abs_avg = torch.mean(torch.abs(self.values), dim=-2)
-        mask = torch.abs(y_std / y_abs_avg) > self.std_tol
-        if not torch.any(mask):
-            raise RuntimeError("All observed points are equal: add more initial samples")
-        # Remove points that have near-null variance: not relevant to the model
-        y_avg = y_avg[mask]
-        y_std = y_std[mask]
-        return mask, y_avg, y_std
-
-    def expand_observations(self, reduced: torch.Tensor) -> torch.Tensor:
-        """Expand reduced observations to full size.
-
-        Parameters
-        ----------
-        reduced : torch.Tensor
-            Reduced observations.
-
-        Returns
[…]
