+++ piglot/optimisers/botorch/bayes.py
@@ -3,7 +3,6 @@ from typing import Tuple, List, Union, Type, Dict, Callable
 from multiprocessing.pool import ThreadPool as Pool
 import os
 import warnings
-import dataclasses
 import numpy as np
 import torch
 from scipy.stats import qmc
@@ -14,6 +13,7 @@ from botorch.optim import optimize_acqf
 from botorch.models import SingleTaskGP
 from botorch.models.model import Model
 from botorch.models.converter import batched_to_model_list
+from botorch.models.transforms import Normalize
 from botorch.acquisition import (
     AcquisitionFunction,
     UpperConfidenceBound,
@@ -36,7 +36,6 @@ from botorch.acquisition.multi_objective import (
 )
 from botorch.utils.multi_objective.box_decompositions.non_dominated import (
     FastNondominatedPartitioning,
-    NondominatedPartitioning,
 )
 from botorch.acquisition.multi_objective.logei import (
     qLogExpectedHypervolumeImprovement,
@@ -190,15 +189,17 @@ class BayesianBoTorch(Optimiser):
             Objective to optimise
         """
 
-    def _build_model(self, std_dataset: BayesDataset) -> Model:
-        # Fetch data
-        params = std_dataset.params
-        values = std_dataset.values
-        variances = std_dataset.variances
-        # Clamp variances to prevent warnings from GPyTorch
+    def _build_model(self, dataset: BayesDataset) -> Model:
+        # Transform outcomes and clamp variances to prevent warnings from GPyTorch
+        values, variances = dataset.transform_outcomes(dataset.values, dataset.variances)
         variances = torch.clamp_min(variances, 1e-6)
         # Initialise model instance depending on noise setting
-        model = SingleTaskGP(params, values, train_Yvar=None if self.noisy else variances)
+        model = SingleTaskGP(
+            dataset.params,
+            values,
+            train_Yvar=None if self.noisy else variances,
+            input_transform=Normalize(d=dataset.params.shape[-1]),
+        )
         # Fit the GP (in case of trouble, fall back to an Adam-based optimiser)
         mll = ExactMarginalLogLikelihood(model.likelihood, model)
         try:
@@ -211,24 +212,23 @@ class BayesianBoTorch(Optimiser):
     def _get_candidates(
             self,
             n_dim,
+            bounds: np.ndarray,
             dataset: BayesDataset,
             test_dataset: BayesDataset,
             ) -> Tuple[np.ndarray, float]:
 
-        # Build model on unit-cube and standardised data
-        std_dataset = dataset.standardised()
-        model = self._build_model(std_dataset)
+        # Build model
+        model = self._build_model(dataset)
 
         # Evaluate GP performance with the test dataset
         cv_error = None
         if self.n_test > 0:
-            std_test_params = test_dataset.normalise(test_dataset.params)
-            std_test_values, _ = dataset.standardise(
+            std_test_values, _ = dataset.transform_outcomes(
                 test_dataset.values,
                 test_dataset.variances,
             )
             with torch.no_grad():
-                posterior = model.posterior(std_test_params)
+                posterior = model.posterior(test_dataset.params)
                 cv_error = (posterior.mean - std_test_values).square().mean().item()
 
         # Build the acquisition function
@@ -237,7 +237,7 @@ class BayesianBoTorch(Optimiser):
         # Optimise acquisition to find next candidate(s)
         candidates, _ = optimize_acqf(
             acq,
-            bounds=torch.stack((std_dataset.lbounds, std_dataset.ubounds)),
+            bounds=torch.from_numpy(bounds.T).to(self.device).to(dataset.dtype),
             q=self.q,
             num_restarts=num_restarts,
             raw_samples=raw_samples,
@@ -247,9 +247,6 @@ class BayesianBoTorch(Optimiser):
             },
         )
 
[…]
