+++ test/examples/dummy_composite_ei.yaml
@@ -3,7 +3,7 @@ iters: 10
 
 optimiser:
   name: botorch
-  acquisition: ei
+  acquisition: qei
 
 
 parameters:
