### 2024-04-17 1643d2a
Add heuristic for reference point and allow manual spec
