+++ piglot/optimisers/botorch/bayes.py
@@ -148,6 +148,7 @@ class BayesianBoTorch(Optimiser):
         load_file: str = None,
         export: str = None,
         device: str = 'cpu',
+        reference_point: List[float] = None,
     ) -> None:
         if not isinstance(objective, GenericObjective):
             raise RuntimeError("Bayesian optimiser requires a GenericObjective")
@@ -166,7 +167,8 @@ class BayesianBoTorch(Optimiser):
         self.n_test = n_test
         self.device = device
         self.partitioning: FastNondominatedPartitioning = None
-        self.ref_point: torch.Tensor = None
+        self.adjusted_ref_point = reference_point is None
+        self.ref_point = None if reference_point is None else -torch.tensor(reference_point)
         if acquisition is None:
             self.acquisition = default_acquisition(
                 objective.composition,
@@ -301,10 +303,11 @@ class BayesianBoTorch(Optimiser):
         return -self.objective.composition.composition_torch(vals, params)
 
     def _update_mo_data(self, dataset: BayesDataset) -> float:
-        # Compute reference point
-        # TODO: check how to properly handle the ref_point and best_value
         y_points = self._composition(dataset.values, dataset.params)
-        self.ref_point = torch.min(y_points, dim=0).values
+        # Compute reference point if needed
+        if self.adjusted_ref_point:
+            nadir = torch.min(y_points, dim=0).values
+            self.ref_point = nadir - 0.1 * (torch.max(y_points, dim=0).values - nadir)
         # Update partitioning and Pareto front
         self.partitioning = FastNondominatedPartitioning(self.ref_point, Y=y_points)
         hypervolume = self.partitioning.compute_hypervolume().item()
