### 2024-04-18 34823af
Retire parameter normalisation and add output of the parameters in the Pareto front
