+++ piglot/parameter.py
@@ -1,5 +1,5 @@
 """Optimisation parameter module."""
-from typing import Iterator, Dict, Any
+from typing import Iterator, Dict, Any, List, Callable
 import os
 from hashlib import sha256
 import numpy as np
@@ -11,19 +11,19 @@ from piglot.utils.yaml_parser import parse_config_file
 class Parameter:
     """Base parameter class."""
 
-    def __init__(self, name, inital_value, lbound, ubound):
+    def __init__(self, name: str, inital_value: float, lbound: float, ubound: float):
         """Constructor for the parameter class.
 
         Parameters
         ----------
         name : str
-            Parameter name
+            Parameter name.
         inital_value : float
-            Initial value for the parameter
+            Initial value for the parameter.
         lbound : float
-            Lower bound for the parameter
+            Lower bound for the parameter.
         ubound : float
-            Upper bound for the parameter
+            Upper bound for the parameter.
         """
         self.name = name
         self.inital_value = inital_value
@@ -32,37 +32,7 @@ class Parameter:
         if inital_value > ubound or inital_value < lbound:
             raise RuntimeError("Initial shot outside of bounds")
 
-    def normalise(self, value):
-        """Normalise a value to internal [-1,1] bounds.
-
-        Parameters
-        ----------
-        value : float
-            Value to normalise
-
-        Returns
-        -------
-        float
-            Normalised value
-        """
-        return 2 * (value - self.lbound) / (self.ubound - self.lbound) - 1
-
-    def denormalise(self, value):
-        """Denormalise a value from internal [-1,1] bounds.
-
-        Parameters
-        ----------
-        value : float
-            Value to denormalise
-
-        Returns
-        -------
-        float
-            Denormalised value
-        """
-        return (self.ubound - self.lbound) * (value + 1) / 2 + self.lbound
-
-    def clip(self, value):
+    def clip(self, value: float) -> float:
         """Clamp a value to the [lbound,ubound] interval.
 
         Parameters
@@ -81,14 +51,14 @@ class Parameter:
 class OutputParameter:
     """Base class for output parameters."""
 
-    def __init__(self, name, mapping):
+    def __init__(self, name: str, mapping: Callable[[Dict[str, float]], float]):
         """Constructor for an output parameter
 
         Parameters
         ----------
         name : str
             Output parameter name
-        mapping : callable
+        mapping : Callable[..., float]
             Function to map from internal to output parameters.
             The internal parameters are passed in an expanded dict.
         """
@@ -105,21 +75,41 @@ class ParameterSet:
 
     def __init__(self):
         """Constructor for a parameter set."""
-        self.parameters = []
+        self.parameters: List[Parameter] = []
 
     def __iter__(self) -> Iterator[Parameter]:
-        """Iterator for a parameter set."""
+        """Iterator for a parameter set.
+
+        Returns
+        -------
+        Iterator[Parameter]
+            Iterator for a parameter set."""
         return iter(self.parameters)
 
     def __len__(self) -> int:
-        """Length of the parameter set."""
+        """Length of the parameter set.
+
+        Returns
+        -------
+        int
+            Length of the parameter set."""
         return len(self.parameters)
 
-    def __getitem__(self, key) -> Parameter:
-        """Get a parameter by name."""
+    def __getitem__(self, key: int) -> Parameter:
+        """Get a parameter by index.
+
+        Parameters
+        ----------
+        key : int
+            Index of the parameter.
+
+        Returns
+        -------
+        Parameter
+            Parameter with the given index."""
         return self.parameters[key]
 
-    def add(self, name, inital_value, lbound, ubound):
+    def add(self, name: str, inital_value: float, lbound: float, ubound: float) -> None:
         """Add a parameter to this set.
 
         Parameters
@@ -127,11 +117,11 @@ class ParameterSet:
[…]
