+++ piglot/solver/solver.py
@@ -335,8 +335,8 @@ class Solver(ABC):
             file.write(f'{result.begin_time - self.begin_time:>15.8e}\t')
             file.write(f'{result.run_time:>15.8e}\t')
             file.write(f'{result.success:>10}\t')
-            for i, param in enumerate(self.parameters):
-                file.write(f"{param.denormalise(result.values[i]):>15.6f}\t")
+            for val in result.values:
+                file.write(f'{val:>15.6f}\t')
             file.write(f'{param_hash}\n')
 
     def get_output_fields(self) -> Dict[str, Tuple[Case, OutputField]]:

[…]
