### 2024-04-23 197bb84
Add option to specify the scale factor for the nadir point
