### 2024-04-23 0c130c4
Add option to skip the evaluation of the initial shot
