### 2024-04-24 c5950e5
Write out internal parameters to CaseResults
