### 2024-04-24 861feb5
Allow to read objective child classes for the design objective
