+++ piglot/solver/solver.py
@@ -357,6 +357,26 @@ class Solver(ABC):
                 output_fields[name] = (case, field)
         return output_fields
 
+    def get_case_params(self, param_hash: str) -> Dict[str, float]:
+        """Get the parameters for a given hash.
+
+        Parameters
+        ----------
+        param_hash : str
+            Hash of the case to load.
+
+        Returns
+        -------
+        Dict[str, float]
+            Parameters for this hash.
+        """
+        case = next(iter(self.cases))
+        result = CaseResult.read(
+            os.path.join(self.cases_hist, f'{case.name()}-{param_hash}'),
+            self.parameters,
+        )
+        return {param.name: result.values[i] for i, param in enumerate(self.parameters)}
+
     def get_output_response(self, param_hash: str) -> Dict[str, OutputResult]:
         """Get the responses from all output fields for a given case.
 
