### 2024-05-02 73147d0
Ignore PCA for non-composite single-objective optimisation
