### 2024-05-03 05c113f
Fix bad device for response weights
