### 2024-05-03 37e91b1
Add error for incompatible acquisitions
