+++ piglot/optimisers/botorch/dataset.py
@@ -5,6 +5,158 @@ import copy
 import numpy as np
 import torch
 
+class Standardiser:
+    """Standardisation transformation."""
+
+    def __init__(self, std_tol: float = 1e-6) -> None:
+        self.mean: torch.Tensor = None
+        self.stds: torch.Tensor = None
+        self.mask: torch.Tensor = None
+        self.std_tol = std_tol
+
+    def fit(self, data: torch.Tensor) -> None:
+        """Fit the standardisation transformation to the given data.
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Data to fit the standardisation transformation to.
+        """
+        self.mean = torch.mean(data, dim=-2)
+        self.stds = torch.std(data, dim=-2)
+        y_abs_avg = torch.mean(torch.abs(data), dim=-2)
+        self.mask = torch.abs(self.stds / y_abs_avg) > self.std_tol
+
+    def transform(
+        self,
+        values: torch.Tensor,
+        variances: torch.Tensor | None = None,
+    ) -> torch.Tensor | Tuple[torch.Tensor, torch.Tensor]:
+        """Standardise data.
+
+        Parameters
+        ----------
+        values : torch.Tensor
+            Values to transform.
+        variances : torch.Tensor | None
+            Variances to transform, if any.
+
+        Returns
+        -------
+        torch.Tensor | Tuple[torch.Tensor, torch.Tensor]
+            Transformed values and variances (if any).
+        """
+        means = (values[:, self.mask] - self.mean[self.mask]) / self.stds[self.mask]
+        if variances is None:
+            return means
+        stds = variances[:, self.mask] / self.stds[self.mask]
+        return means, stds
+
+    def untransform(self, data: torch.Tensor) -> torch.Tensor:
+        """Unstandardise data.
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Data to unstandardise.
+
+        Returns
+        -------
+        torch.Tensor
+            Unstandardised data.
+        """
+        values = data * self.stds[self.mask] + self.mean[self.mask]
+        # Nothing more to do if no outputs are suppressed
+        if torch.all(self.mask):
+            return values
+        # Infer the shape of the expanded tensor: only modify the last dimension
+        new_shape = list(values.shape)
+        new_shape[-1] = self.mean.numel()
+        expanded = torch.empty(new_shape, dtype=values.dtype, device=values.device)
+        # Fill the tensor using a 2D view:
+        # i) modelled outcomes are directly inserted
+        # ii) missing outcomes are filled with the average of the observed ones
+        expanded_flat = expanded.view(-1, new_shape[-1])
+        expanded_flat[:, self.mask] = values.view(-1, values.shape[-1])
+        expanded_flat[:, ~self.mask] = self.mean[~self.mask]
+        # Note: we are using a view, so the expanded tensor is already modified
+        return expanded
+
+
+class PCA:
+    """Principal Component Analysis transformation."""
+
+    def __init__(self, variance: float) -> None:
+        self.variance = variance
+        self.standardiser = Standardiser()
+        self.num_components: int = None
+        self.transformation: torch.Tensor = None
+
+    def fit(self, data: torch.Tensor) -> None:
+        """Fit the PCA transformation to the given data.
+
+        Parameters
+        ----------
+        data : torch.Tensor
+            Data to fit the PCA transformation to.
+        """
+        # Standardise data
+        self.standardiser.fit(data)
+        data_std: torch.Tensor = self.standardiser.transform(data)
+        # Compute eigenvalues and vectors of the covariance matrix and sort by decreasing variance
+        vals, vecs = torch.linalg.eigh(torch.cov(data_std.T))
+        idx = torch.argsort(vals, descending=True)
+        vals = vals[idx]
+        vecs = vecs[:, idx]
[…]
