### 2024-05-06 28bface
Sanitise transformer keywords
