### 2024-05-13 9ab1b1d
Add test for sane reduction functions
