### 2024-05-13 95cf3da
Add more exceptions if the reduction fails
