### 2024-05-21 20df61e
Limit initial batch size for raw samples
