### 2024-05-21 0f934a8
Fix bad device in EndpointFlattenUtility
