### 2024-05-21 925e9db
Add option to specify default torch device
