### 2024-05-24 f37a37d
Robustify `interpolate_response` for scalar responses
