### 2024-05-24 20b7404
Restore check for equal observed points
