### 2024-05-27 5762e47
Add support for response transformers for design and fitting objectives
