+++ piglot/utils/response_transformer.py
@@ -1,0 +1,202 @@
+"""Module for defining transformations for responses."""
+from typing import Union, Dict, Any, Type, List
+from abc import ABC, abstractmethod
+import numpy as np
+from piglot.utils.assorted import read_custom_module
+from piglot.solver.solver import OutputResult
+
+
+class ResponseTransformer(ABC):
+    """Abstract class for defining transformation functions."""
+
+    @abstractmethod
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        OutputResult
+            Transformed time and data points of the response.
+        """
+
+
+class ChainResponse(ResponseTransformer):
+    """Chain of response transformers."""
+
+    def __init__(self, transformers: List[Any]):
+        self.transformers = [read_response_transformer(t) for t in transformers]
+
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        Output
+            Transformed time and data points of the response.
+        """
+        for transformer in self.transformers:
+            response = transformer.transform(response)
+        return response
+
+
+class MinimumResponse(ResponseTransformer):
+    """Minimum of a response transformer."""
+
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        OutputResult
+            Transformed time and data points of the response.
+        """
+        return OutputResult(np.array([0.0]), np.array([np.min(response.data)]))
+
+
+class MaximumResponse(ResponseTransformer):
+    """Maximum of a response transformer."""
+
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        OutputResult
+            Transformed time and data points of the response.
+        """
+        return OutputResult(np.array([0.0]), np.array([np.max(response.data)]))
+
+
+class NegateResponse(ResponseTransformer):
+    """Negate a response transformer."""
+
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        OutputResult
+            Transformed time and data points of the response.
+        """
+        return OutputResult(response.time, -response.data)
+
+
+class SquareResponse(ResponseTransformer):
+    """Square a response transformer."""
+
+    def transform(self, response: OutputResult) -> OutputResult:
+        """Transform the input data.
+
+        Parameters
+        ----------
+        response : OutputResult
+            Time and data points of the response.
+
+        Returns
+        -------
+        Output
+            Transformed time and data points of the response.
+        """
+        return OutputResult(response.time, np.square(response.data))
+
+
+class AffineTransformResponse(ResponseTransformer):
+    """Affine transformation of a response transformer."""
+
+    def __init__(
+        self,
+        scale_x: float = 1.0,
+        offset_x: float = 0.0,
+        scale_y: float = 1.0,
+        offset_y: float = 0.0,
+    ):
+        self.scale_x = scale_x
+        self.offset_x = offset_x
[…]

[…]
