### 2024-05-28 10cf61d
Fix on abaqus fields check
