### 2024-05-28 6ed1385
Optimised the reading of fields and fixed some bugs
