+++ piglot/solver/abaqus/fields.py
@@ -111,8 +111,10 @@ class AbaqusInputData(InputData):
             self.job_name = self.__sanitize_field(self.job_name, job_list, "job")
 
             instance_list = re.findall(r'\*Instance, name=([^,]+)', data)
-            self.instance_name = self.__sanitize_field(self.instance_name, instance_list,
-                                                       "instance")
+            self.instance_name = self.__sanitize_field(self.instance_name.upper(),
+                                                       instance_list,
+                                                       "instance"
+                                                    )
 
             step_list = re.findall(r'\*Step, name=([^,]+)', data)
             self.step_name = self.__sanitize_field(self.step_name, step_list, "step")
