### 2024-05-28 9b4883a
Deleted unnecessary code to delete temporary files that are not being generated anymore
