### 2024-05-28 7294ae7
Optimised the writing of the field parameters
