### 2024-05-28 6e6472d
Edited the raise value error output
