+++ piglot/solver/abaqus/solver.py
@@ -143,7 +143,6 @@ class AbaqusSolver(Solver):
             cwd=tmp_dir,
             shell=False,
             stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
             check=False
         )
         end_time = time.time()
