### 2024-05-28 527cf3e
Fixed multiple spaces before operator
