### 2024-05-29 09b30e3
Rewrited the reader.py code to make the methods less complex.
