+++ piglot/solver/abaqus/reader.py
@@ -18,29 +18,17 @@ def input_variables():
     args = sys.argv
     variables = {}
 
-    input_file_list = [a for a in args if a.startswith("input_file=")]
-    job_name_list = [a for a in args if a.startswith("job_name=")]
-    step_name_list = [a for a in args if a.startswith("step_name=")]
-    instance_name_list = [a for a in args if a.startswith("instance_name=")]
-    set_name_list = [a for a in args if a.startswith("set_name=")]
-    field_list = [a for a in args if a.startswith("field=")]
-    x_field_list = [a for a in args if a.startswith("x_field=")]
+    variable_names = ['input_file',
+                      'job_name',
+                      'step_name',
+                      'instance_name',
+                      'set_name',
+                      'field',
+                      'x_field']
 
-    # Checks if the input_file, job_name, step_name and instance_name are not empty
-    variables['input_file'] = input_file_list[0].replace('input_file=', '') \
-        if input_file_list else None
-    variables['job_name'] = job_name_list[0].replace('job_name=', '') \
-        if job_name_list else None
-    variables['step_name'] = step_name_list[0].replace('step_name=', '') \
-        if step_name_list else None
-    variables['instance_name'] = instance_name_list[0].replace('instance_name=', '') \
-        if instance_name_list else None
-    variables['set_name'] = set_name_list[0].replace('set_name=', '') \
-        if set_name_list else None
-    variables['field'] = field_list[0].replace('field=', '') \
-        if field_list else None
-    variables['x_field'] = x_field_list[0].replace('x_field=', '') \
-        if x_field_list else None
+    for var_name in variable_names:
+        var_list = [a for a in args if a.startswith(f"{var_name}=")]
+        variables[var_name] = var_list[0].replace(f'{var_name}=', '') if var_list else None
 
     return variables
 
@@ -67,12 +55,16 @@ def file_name_func(set_name, variable_name, inp_name):
 
 
 def field_location(i, variables_array, output_variable, location):
-    """It gets the node data of the specified node set.
+    """It gets the node data of the specified node set. Create a variable that refers to the output 
+    variable of the node set. If the field is S or E it extrapolates the data to the nodes, if the 
+    field is U or RF the data is already on the nodes so it doesn't need extrapolation.
 
     Parameters
     ----------
     i : int
-        It is an int number, 0 or 1 in the case of the stresses and strains.
+        It is an int number that represents the index of the variables_array.
+    variables_array : list
+        It is a list that contains the field variables (S, U, RF, E or LE)
     output_variable : str
         Its the output variable (S, U, RF, E or LE) of the nodes.
     location : str
@@ -92,106 +84,146 @@ def field_location(i, variables_array, output_variable, location):
     return location_output_variable
 
 
-def main():
-    """Main function of the reader.py
-    """
+def get_nlgeom_setting(inp_name):
+    """It verifies if the 'nlgeom' setting is set to 'YES' in the input file.
 
+    Parameters
+    ----------
+    inp_name : str
+        The name of the input file.
+
+    Returns
+    -------
+    int
+        It returns 1 if the 'nlgeom' setting is set to 'YES' in the input file, otherwise it 
+        returns 0.
+
+    Raises
+    ------
+    ValueError
+        Raises an error if the 'nlgeom' setting is not found in the input file.
+    """
+    with codecs.open(inp_name, 'r', encoding='utf-8') as input_file:
+        file_content = input_file.read()
+        match = re.search(r'\*Step.*nlgeom=(\w+)', file_content)
+        if match:
+            nlgeom_setting = match.group(1)
[…]
