### 2024-05-29 3503ac2
removed fstrings from reader.py
