### 2024-05-29 fd5240d
fixed a bug in the write_output_file method
