### 2024-05-29 ffdf9ee
Removed redundant variable
