### 2024-05-29 3addfdd
First version of solver output streams
