### 2024-05-30 9e82563
Added the suppresion of errors from abaqus
