### 2024-05-30 a4458fa
Optimised Abaqus interface (#24)
- Optimised the way the field outputs are written and treated inside
piglot. Now piglot only writes the fields and the set names that are
present in the YAML file and automatically extrapolates to nodes units
that are calculated at the integration points;
- Added errors to check for non-linear fields in linear simulations;
- Fixed some bugs on the implementation of the selected instance on
abaqus;
- Generalised the reading of Abaqus input fiiles.
