### 2024-05-31 3aef57c
Removed job_name from the input file and make the comparison case insensitive.
