+++ piglot/solver/abaqus/fields.py
@@ -13,12 +13,18 @@ from piglot.utils.solver_utils import write_parameters, has_parameter, get_case_
 class AbaqusInputData(InputData):
     """Container for Abaqus input data."""
 
-    def __init__(self, input_file: str, job_name: str, step_name: str, instance_name: str) -> None:
+    def __init__(
+        self,
+        input_file: str,
+        step_name: str,
+        instance_name: str,
+        job_name: str = None,
+    ) -> None:
         super().__init__()
         self.input_file = input_file
-        self.job_name = job_name
         self.step_name = step_name
         self.instance_name = instance_name
+        self.job_name = job_name
 
     def prepare(
             self,
@@ -42,10 +48,20 @@ class AbaqusInputData(InputData):
         AbaqusInputData
             Input data prepared for the simulation.
         """
+        # Sanitize job_name
+        if self.job_name is None:
+            raise ValueError("Job name not specified. Call the 'check' method first.")
+
         # Copy file and write out the parameters
         dest_file = os.path.join(tmp_dir, os.path.basename(self.input_file))
         write_parameters(parameters.to_dict(values), self.input_file, dest_file)
-        return AbaqusInputData(dest_file, self.job_name, self.step_name, self.instance_name)
+
+        return AbaqusInputData(
+            dest_file,
+            self.step_name,
+            self.instance_name,
+            job_name=self.job_name,
+        )
 
     @staticmethod
     def __sanitize_field(field_name: str, field_list: List[str], keyword: str) -> str:
@@ -106,7 +122,7 @@ class AbaqusInputData(InputData):
             data = file.read()
 
             job_list = re.findall(r'\*\* Job name: ([^M]+)', data)
-            job_list = [job.strip() for job in job_list]  # Remove trailing whitespace
+            job_list = [job.strip() for job in job_list]
             self.job_name = self.__sanitize_field(self.job_name, job_list, "job")
 
             instance_list = re.findall(r'\*Instance, name=([^,]+)', data)
@@ -140,8 +156,7 @@ class AbaqusInputData(InputData):
         AbaqusInputData
             Current input data.
         """
-        return AbaqusInputData(os.path.join(target_dir, os.path.basename(self.input_file)),
-                               self.job_name, self.step_name, self.instance_name)
+        raise RuntimeError("Abaqus does not support extracting current results.")
 
 
 class FieldsOutput(OutputField):
