### 2024-06-02 #6652 @pprachas
Hi @ruicoelhopedro,

First of all, sorry for the late review. I am very impressed by the work done here -- this package will be very useful in the scientific community! I was able to quickly install package (both using pip and pipx) and run example programs both from terminal and python scripts quickly. Here are a few suggestions I have:

1. I think the first example shown here is a good example -- it's simple enough to understand and its easy to see that the software is working. However, its not showing all the capabilities of the solver especially the output. For example, showing the convergence history (which I found digging around other in examples) in the first example would be helpful since convergence is an important especially during troubleshooting. I think having the first example show off all the important capabilities of the package or list all the possible capabilities and linking them to examples will be extremely helpful to users.

2. Can the output images also be saved like the gifs? It might also be helpful to explain the location and name of the output files/gifs/images for clarity.

3. Is there a general guide to integrate other custom solvers and objective functions in general? For example, If I have a custom solver in Matlab/Python or a non-standard objective function, how do I use piglot in this case?

Looking forward to your response!
