### 2024-06-04 4d50458
Added a flag to enable the reading of strings with spaces in the input file
