### 2024-06-04 5925eb5
Minor fixes in piglot-plot
