+++ docs/source/examples/post_processing/description.md
@@ -1,0 +1,273 @@
+# Post-processing results
+
+When solving a problem with a YAML configuration file, like the one in the [simple curve fitting example](../sample_curve_fitting/description.md), `piglot` stores most of the data collected during the optimisation in the output directory.
+In this example, we will walk over the structure of the output data and, particularly, on using our `piglot-plot` utility for handling the post-processing and visualisation.
+We build on top of the [simple curve fitting example](../sample_curve_fitting/description.md) and we assume in this tutorial that the optimisation problem has already been solved with
+```bash
+piglot config.yaml
+```
+which produces the `config` output directory.
+
+
+## Output directory
+
+When running `piglot config.yaml`, the following directory structure will be produced:
+```
+config
+├── cases
+│   └── case_1
+├── cases_hist
+│   ├── case_1-066c8026ba74b70857f3ab68b24a4c1200e3eecf472745a3172372e57a087d79
+│   ├── case_1-156da358cdfe9167d0eed7b536182d94fe76ee8572e98e1008b559850a26d94a
+|   |   ...
+│   └── case_1-f096b92e4c546d96828afd012c148cbb70a80c93005ee4cf08e7c31d8e2e30a0
+├── config
+├── func_calls
+├── history
+├── progress
+└── tmp
+    ├── case_1.dat
+    ├── case_1.out
+    └── case_1.tmp
+config.yaml
+reference_curve.txt
+```
+Inside the `config` directory, three directories (`cases`, `cases_hist` and `tmp`) and four files (`config`, `func_calls`, `history` and `progress`) have been created.
+Each of the files has the following data:
+- `config/config` is simply a copy of the parsed YAML input file, used to verify that the input reading is correct and to store the exact configuration of the optimisation case;
+- `config/func_calls` contains tabular data for each function evaluation, namely the start and objective evaluation run times, the objective value (and possibly its variance in [stochastic optimisation](../sample_curve_fitting_stochastic/description.md)), the parameter values and the parameter hash.
+  The hash is computed with SHA256 on the binary representation of the parameters and serves as the unique identifier of the case.
+  An example of this file is:
+  ```
+    Start Time /s	    Run Time /s	      Objective	              a	                                                            Hash
+   1.33138150e-03	 1.10529615e-02	 4.32774678e-01	       1.000000	e1c54f41b449d2997ce426b22b0e24103c258a4e35632dcce8da80d964140bd8
+   2.12348597e-02	 7.60192145e-03	 8.22040327e-01	       0.621861	e440d899a5137f4cf4559e98a15418b8e9c783be1b1ce8549f3e513cc06a816a
+   2.88948463e-02	 7.28006847e-03	 1.04827989e-01	       2.491877	64d50061013020fdda2cbe15f29510bd9a0cc952726bb577db71179a639c4359
+   3.62300798e-02	 7.32552260e-03	 1.59057447e+00	       3.916550	4a3b2ea1387fac34c7ae1a113cf680eac6ea483c8819da33d5ad53db26b2dd3f
+   4.36127670e-02	 7.15153851e-03	 4.04572983e-01	       1.033125	bfa9fb1b7f2f2e46549ca9bd8c00bee8369e57201a70cbdbd272406cb9bd58c9
+   5.08190878e-02	 6.92950096e-03	 1.38579282e-02	       1.820899	5a50cd302ecd881a9fe5fb6b75f5315780d2d063748693ff2b0d223aea163ae1
+   5.78034185e-02	 6.93223719e-03	 6.14725627e-01	       3.191400	743b10cb79270796d1257cfbd2790a8dc2dba58b3b724af594dba0554af76a8a
[…]
