+++ docs/source/index.rst
@@ -106,6 +106,7 @@ If you wish to install all optimisers at once, you can run :code:`pip install pi
    examples/fitting
    examples/design
    examples/analytical
+   examples/post_processing/description
 
 
 .. toctree::
