### 2024-06-05 f332e12
Add guide for implementing new solvers
