### 2024-06-05 4d65124
Attempt to fix the JOSS paper YAML
