+++ .github/workflows/generate_pdf.yaml
@@ -9,6 +9,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+      - name: Install dependencies
+        run: sudo apt-get install -y librsvg2-bin
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
