### 2024-06-05 #6652 @ruicoelhopedro
> Hi @ruicoelhopedro,
>
> First of all, sorry for the late review. I am very impressed by the work done here -- this package will be very useful in the scientific community! I was able to quickly install package (both using pip and pipx) and run example programs both from terminal and python scripts quickly. Here are a few suggestions I have:

Thank you for the feedback! We are happy that everything is working smoothly!
Meanwhile, we have started this PR (CM2S/piglot#26) for the review, where we've already added your suggestions to it.

> 1. I think the first example shown here is a good example -- it's simple enough to understand and its easy to see that the software is working. However, its not showing all the capabilities of the solver especially the output. For example, showing the convergence history (which I found digging around other in examples) in the first example would be helpful since convergence is an important especially during troubleshooting. I think having the first example show off all the important capabilities of the package or list all the possible capabilities and linking them to examples will be extremely helpful to users.

Thank you for the response! Our first example is mostly focused on the main steps for solving a problem, but we agree that the post-processing features should be exposed in more detail. In this line, we updated the documentation as follows:
- Added a new dedicated example for showcasing the features in `piglot-plot`. While there, we also fixed some wrong arguments in it.
- Updated the first example to include the list of all features supported by `piglot-plot` with appropriate links to the new example and added a section showing the convergence history.

> 2. Can the output images also be saved like the gifs? It might also be helpful to explain the location and name of the output files/gifs/images for clarity.

Yes, most plotting options support saving with the `--save_fig` option. This has been extensively addressed in the new example.

> 3. Is there a general guide to integrate other custom solvers and objective functions in general? For example, If I have a custom solver in Matlab/Python or a non-standard objective function, how do I use piglot in this case?

Thank you for bringing up this point, this was lacking. Currently, the interface is tailored for the setup with FEM codes using input files and, as a consequence, is quite verbose and complex. We aim to drastically simplify it in the upcoming months. Nevertheless, we added a page to the documentation describing the main steps for implementing a new solver interface, along with a simple example.
Regarding a non-standard objective function, with the current main version, we don't explicitly support custom objectives; these can be introduced, but require re-implementing most of the objective code from scratch. We are currently working on a large refactoring in the development branch that will greatly simplify this topic and, importantly, allow the user to introduce custom objectives, reductions, quantities and solvers via external scripts to `piglot`. There are some congresses and deadlines in the upcoming month, but afterwards, these features will quickly make their way to main.
