### 2024-06-06 a39c7a6
Add comment on save_fig format
