### 2024-06-06 804d1c8
Fix tests for `piglot-plot`
