### 2024-06-07 288da13
Generalised the reading for any Abaqus version. (#25)
- Removed the `job_name` from the `.yaml` file since a Abaqus input file
only has one job;
- Made the reading and comparison with Abaqus insensitive to upper and
lower cases;
- Added a flag to support the reading of strings with spaces in the
`.yaml` file and successfully find them in Abaqus `.inp` file.
