+++ examples/abaqus_solver_fitting/config.yaml
@@ -19,7 +19,6 @@ objective:
 
     cases:
       'sample.inp':
-        job_name: Job-1 # optional field for this case
         step_name: Step-1 # optional field for this case
         instance_name: Part-1-1 # optional field for this case
         fields:
