### 2024-06-07 a97f0d4
Fix bad ModelList when there is a single output in MOBO
