### 2024-06-09 #6652 @pprachas
Hi @hugoledoux ,

The changes look great! As for my last point about objective functions, I understand that this use case might be more niche and I appreciate the efforts to try to accommodate this case.

@hugoledoux `piglot` is nice open-source optimizer that covers a large range of common optimization problems in science and engineering. I have verified that example problems can be reproduced with minimal effort and the package is extremely well documented. I believe that the changes the developers proposed to accommodate more niche cases will be implements in due time and will elevate the package even further, but I am happy to recommend the package  for publication in JOSS in its current state.
