### 2024-06-12 8b0ff6c
Deleted f-string (#27)
Deleted f-string in python 2.7 script that was preventing the file
"reader.py" to run properly due to sintax error.
