### 2024-06-17 #6652 @Extraweich
Hello @ruicoelhopedro,

please excuse my late response in the reviewing process. I have just started reviewing and immediately think it is a great project. Especially the integration of the Abaqus solver I find very interesting.
As a side note, I was wondering if there is a way to integrate the communication with a workload management tool (e.g. SLURM) instead of running Abaqus locally. The idea comes from the background that most research institutions will use Abaqus on a cluster based on shared licenses and that it is a bad practice to use these licenses locally or start a simulation without queuing the job accordingly with the implemented workload management tool.

Other than that, I think that @pprachas has already provided valuable input, which you have addressed accordingly.

I'll take my time to finish the review process soon and you'll hear from me if I have further suggestions for any adaptations.
