### 2024-06-17 fc57b46
Add utility for training a surrogate model on the data
