### 2024-06-17 3a4c6ad
Introduce a query optimiser
