### 2024-06-18 #6652 @ruicoelhopedro
> Hello @ruicoelhopedro,
>
> please excuse my late response in the reviewing process. I have just started reviewing and immediately think it is a great project. Especially the integration of the Abaqus solver I find very interesting. As a side note, I was wondering if there is a way to integrate the communication with a workload management tool (e.g. SLURM) instead of running Abaqus locally. The idea comes from the background that most research institutions will use Abaqus on a cluster based on shared licenses and that it is a bad practice to use these licenses locally or start a simulation without queuing the job accordingly with the implemented workload management tool.
>
> Other than that, I think that @pprachas has already provided valuable input, which you have addressed accordingly.
>
> I'll take my time to finish the review process soon and you'll hear from me if I have further suggestions for any adaptations.

Thank you for your input!

The remote runner is a very interesting application we did not consider, and we deeply appreciate the suggestion. In the current state, this is not supported; `piglot` assumes the run occurs locally. Our experience with workload managers such as SLURM is quite limited. Nonetheless, the problem seems strongly related to allowing running the solver via SSH and SFTP on a remote machine, which may be a good starting point for implementing this feature. From a top-level perspective, two candidate approaches come to mind:
- If `piglot` can be installed on the runner, then we can evaluate the solver calls using a wrapper that runs individual simulations locally on the runner (and thus does not require any modification to the solver interface). This also requires a wrapper for sending and receiving data between the host and the runner, which can be implemented as some sort of `remote` solver. Similarly, support for SLURM can be added as another layer of this abstraction. All existing solvers should work out of the box with this approach.
- If `piglot` is not available on the runner, then this becomes more complicated, as either:
  - Support for sending/receiving data is implemented per solver, which may require extensive difficult-to-debug changes;
  - Force tunneling all the interactions with the filesystem and OS via SSH, which again requires adapting existing solvers and may drastically increase complexity.

Once again, we appreciate the suggestion, and it has been added to our TODO list. Likely the first option is the most sensible approach, but if the installation of `piglot` in the worker becomes too restrictive, then the second option will likely have to be considered. We are looking forward to your opinion on the topic!
