+++ piglot/optimisers/query.py
@@ -1,6 +1,11 @@
 """Module for a simple query optimiser."""
 from typing import Tuple
+import os
 import numpy as np
+import torch
+from botorch.utils.multi_objective.box_decompositions.non_dominated import (
+    FastNondominatedPartitioning,
+)
 from piglot.optimiser import Optimiser, InvalidOptimiserException
 from piglot.objective import Objective, GenericObjective
 
@@ -12,6 +17,8 @@ class QueryOptimiser(Optimiser):
         self,
         objective: Objective,
         param_list_file: str,
+        reference_point: list[float] = None,
+        nadir_scale: float = 0.1,
     ) -> None:
         """Constructor for the Query optimiser class.
 
@@ -21,9 +28,15 @@ class QueryOptimiser(Optimiser):
             Objective to optimise.
         param_list_file : str
             File containing the list of parameters.
+        reference_point : list[float], optional
+            Reference point for multi-objective optimisation, by default None.
+        nadir_scale : float, optional
+            Scale factor for the nadir point, by default 0.1.
         """
         super().__init__('Query', objective)
         self.param_list = np.genfromtxt(param_list_file)
+        self.reference_point = reference_point
+        self.nadir_scale = nadir_scale
         if len(self.param_list.shape) == 1:
             self.param_list = self.param_list.reshape(-1, 1)
 
@@ -41,6 +54,49 @@ class QueryOptimiser(Optimiser):
             raise InvalidOptimiserException('This optimiser does not support composition')
         if objective.stochastic:
             raise InvalidOptimiserException('This optimiser does not support stochasticity')
+
+    def update_mo_data(self, parameters: np.ndarray, observations: np.ndarray) -> Tuple[float, int]:
+        """Get the partitioning of the observations in multi-objective optimisation.
+
+        Parameters
+        ----------
+        parameters : np.ndarray
+            Array of parameters.
+        observations : np.ndarray
+            Array of observations.
+
+        Returns
+        -------
+        Tuple[float, int]
+            Hypervolume and number of non-dominated points.
+        """
+        if self.reference_point is not None:
+            ref_point = np.array(self.reference_point)
+        else:
+            nadir = np.max(observations, axis=0)
+            ref_point = nadir + self.nadir_scale * (nadir - np.min(observations, axis=0))
+        partitioning = FastNondominatedPartitioning(
+            ref_point=torch.from_numpy(-ref_point),
+            Y=torch.from_numpy(-observations),
+        )
+        hypervolume = partitioning.compute_hypervolume().item()
+        pareto = -partitioning.pareto_Y
+        # Map each Pareto point to the original parameter space
+        param_indices = [
+            torch.argmin((torch.from_numpy(observations) - pareto[i, :]).norm(dim=1)).item()
+            for i in range(pareto.shape[0])
+        ]
+        # Dump the Pareto front to a file
+        with open(os.path.join(self.output_dir, "pareto_front"), 'w', encoding='utf8') as file:
+            # Write header
+            num_obj = pareto.shape[1]
+            file.write('\t'.join([f'{"Objective_" + str(i + 1):>15}' for i in range(num_obj)]))
+            file.write('\t' + '\t'.join([f'{param.name:>15}' for param in self.parameters]) + '\n')
+            # Write each point
+            for i, idx in enumerate(param_indices):
+                file.write('\t'.join([f'{-x.item():>15.8f}' for x in pareto[i, :]]) + '\t')
+                file.write('\t'.join([f'{x.item():>15.8f}' for x in parameters[idx, :]]) + '\n')
+        return -np.log(hypervolume)
 
     def _optimise(
         self,
@@ -83,14 +139,35 @@ class QueryOptimiser(Optimiser):
[…]
