### 2024-06-27 CM2S/piglot#28 @Extraweich
piglot\examples\crate_solver_fitting predicted.dat missing
I was not able to run the given example due to a missing predicted.dat.
