### 2024-06-27 CM2S/piglot#29 @Extraweich
piglot\test\test_examples_plots.py not running properly
All tests run as they should, but the given test above results in the following output:

```
============================= test session starts =============================
platform win32 -- Python 3.9.18, pytest-8.1.1, pluggy-1.4.0
rootdir: D:\Code\Python\piglot
configfile: pyproject.toml
plugins: anyio-3.5.0, jaxtyping-0.2.30, typeguard-2.13.3
collected 151 items

test\test_examples.py .................................................. [ 33%]
....                                                                     [ 35%]
test\test_examples_assertions.py ....................................... [ 61%]
                                                                         [ 61%]
test\test_examples_plots.py

Fatal Python error: Aborted

Thread 0x00003af0 (most recent call first):
  File "C:\Anaconda3\lib\threading.py", line 316 in wait
  File "C:\Anaconda3\lib\threading.py", line 581 in wait
  File "C:\Anaconda3\lib\site-packages\tqdm\_monitor.py", line 60 in run
  File "C:\Anaconda3\lib\threading.py", line 980 in _bootstrap_inner
  File "C:\Anaconda3\lib\threading.py", line 937 in _bootstrap

Current thread 0x00006020 (most recent call first):
  File "C:\Anaconda3\lib\site-packages\matplotlib\transforms.py", line 2436 in get_affine
  File "C:\Anaconda3\lib\site-packages\matplotlib\transforms.py", line 2437 in get_affine
  File "C:\Anaconda3\lib\site-packages\matplotlib\patches.py", line 587 in draw
  File "C:\Anaconda3\lib\site-packages\matplotlib\artist.py", line 72 in draw_wrapper
  File "C:\Anaconda3\lib\site-packages\matplotlib\figure.py", line 3153 in draw
  File "C:\Anaconda3\lib\site-packages\matplotlib\artist.py", line 72 in draw_wrapper
  File "C:\Anaconda3\lib\site-packages\matplotlib\artist.py", line 95 in draw_wrapper
  File "C:\Anaconda3\lib\site-packages\matplotlib\backends\backend_agg.py", line 388 in draw
  File "C:\Anaconda3\lib\site-packages\matplotlib\backends\backend_agg.py", line 445 in _print_pil
  File "C:\Anaconda3\lib\site-packages\matplotlib\backends\backend_agg.py", line 497 in print_png
  File "C:\Anaconda3\lib\site-packages\matplotlib\backend_bases.py", line 2043 in <lambda>
  File "C:\Anaconda3\lib\site-packages\matplotlib\backend_bases.py", line 2187 in print_figure
  File "C:\Anaconda3\lib\site-packages\matplotlib\backends\backend_qtagg.py", line 75 in print_figure
  File "C:\Anaconda3\lib\site-packages\matplotlib\figure.py", line 3390 in savefig
  File "C:\Anaconda3\lib\site-packages\piglot\bin\piglot_plot.py", line 112 in plot_best
  File "C:\Anaconda3\lib\site-packages\piglot\bin\piglot_plot.py", line 548 in main
  File "D:\Code\Python\piglot\test\test_examples_plots.py", line 59 in test_input_files
  File "C:\Anaconda3\lib\site-packages\_pytest\python.py", line 195 in pytest_pyfunc_call
  File "C:\Anaconda3\lib\site-packages\pluggy\_callers.py", line 102 in _multicall
  File "C:\Anaconda3\lib\site-packages\pluggy\_manager.py", line 119 in _hookexec
  File "C:\Anaconda3\lib\site-packages\pluggy\_hooks.py", line 501 in __call__
  File "C:\Anaconda3\lib\site-packages\_pytest\python.py", line 1772 in runtest
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 172 in pytest_runtest_call
  File "C:\Anaconda3\lib\site-packages\pluggy\_callers.py", line 102 in _multicall
  File "C:\Anaconda3\lib\site-packages\pluggy\_manager.py", line 119 in _hookexec
  File "C:\Anaconda3\lib\site-packages\pluggy\_hooks.py", line 501 in __call__
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 240 in <lambda>
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 340 in from_call
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 239 in call_and_report
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 134 in runtestprotocol
  File "C:\Anaconda3\lib\site-packages\_pytest\runner.py", line 115 in pytest_runtest_protocol
  File "C:\Anaconda3\lib\site-packages\pluggy\_callers.py", line 102 in _multicall
  File "C:\Anaconda3\lib\site-packages\pluggy\_manager.py", line 119 in _hookexec
  File "C:\Anaconda3\lib\site-packages\pluggy\_hooks.py", line 501 in __call__
  File "C:\Anaconda3\lib\site-packages\_pytest\main.py", line 364 in pytest_runtestloop
  File "C:\Anaconda3\lib\site-packages\pluggy\_callers.py", line 102 in _multicall
  File "C:\Anaconda3\lib\site-packages\pluggy\_manager.py", line 119 in _hookexec
  File "C:\Anaconda3\lib\site-packages\pluggy\_hooks.py", line 501 in __call__
  File "C:\Anaconda3\lib\site-packages\_pytest\main.py", line 339 in _main
  File "C:\Anaconda3\lib\site-packages\_pytest\main.py", line 285 in wrap_session
  File "C:\Anaconda3\lib\site-packages\_pytest\main.py", line 332 in pytest_cmdline_main
  File "C:\Anaconda3\lib\site-packages\pluggy\_callers.py", line 102 in _multicall
  File "C:\Anaconda3\lib\site-packages\pluggy\_manager.py", line 119 in _hookexec
  File "C:\Anaconda3\lib\site-packages\pluggy\_hooks.py", line 501 in __call__
  File "C:\Anaconda3\lib\site-packages\_pytest\config\__init__.py", line 174 in main
  File "C:\Anaconda3\lib\site-packages\_pytest\config\__init__.py", line 197 in console_main
  File "C:\Anaconda3\Scripts\pytest.exe\__main__.py", line 7 in <module>
  File "C:\Anaconda3\lib\runpy.py", line 87 in _run_code
  File "C:\Anaconda3\lib\runpy.py", line 197 in _run_module_as_main
```

This could be a problem with my own installation, though. Maybe worth to look into it.
