### 2024-06-27 #6652 @Extraweich
I've finished the review process. Lovely project with a good documentation!

@hugoledoux, I very much recommend this project to be published in JOSS.

A few minor remarks, which are not necessary to be solved before publishing:

- I found an issue with an [automated test](https://github.com/CM2S/piglot/issues/29), which could be specific to my system
- a single [example ](https://github.com/CM2S/piglot/issues/28) did not run properly, due to a missing file
