### 2024-06-27 a0d8717
Lock numpy<2
From https://github.com/pytorch/pytorch/issues/107302, the .numpy() methods may fail with incompatible versions (and Windows CI was failing)
