+++ pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
     "Programming Language :: Python :: 3 :: Only",
 ]
 dependencies = [
-    "numpy",
+    "numpy<2", # https://github.com/pytorch/pytorch/issues/107302
     "tqdm",
     "pandas",
     "sympy",
