### 2024-07-05 CM2S/piglot#28 @ruicoelhopedro
Thank you for the feedback! This was due to a git ignore entry and has been fixed in #26.
