### 2024-07-05 CM2S/piglot#29 @ruicoelhopedro
Thanks for the report! This seems to be related to a `fig.savefig()` call, which is quite strange. Can you confirm if a MWE like this works in your installation?

```
import numpy as np
import matplotlib.pyplot as plt

x = np.linspace(0, 5, 50)
y = np.sin(x)
fig, ax = plt.subplots()
ax.plot(x, y)
fig.savefig('tmp.png')
```
