### 2024-07-05 #6652 @ruicoelhopedro
Thank you for the feedback and we apologise for the delayed response (due to a conference). Our revision PR (CM2S/piglot#26) has been updated accordingly and should be ready to merge.
