### 2024-07-09 a15ac01
Add BO hyperparams to config and various fixes
