+++ piglot/optimisers/botorch/bayes.py
@@ -149,7 +149,7 @@ class BayesianBoTorch(Optimiser):
     def __init__(
         self,
         objective: Objective,
-        n_initial: int = 8,
+        n_initial: int = None,
         n_test: int = 0,
         acquisition: str = None,
         beta: float = 1.0,
@@ -163,6 +163,12 @@ class BayesianBoTorch(Optimiser):
         nadir_scale: float = 0.1,
         skip_initial: bool = False,
         pca_variance: float = None,
+        num_restarts: int = None,
+        raw_samples: int = None,
+        mc_samples: int = None,
+        batch_size: int = None,
+        num_fantasies: int = None,
+        sequential: bool = False,
     ) -> None:
         if not isinstance(objective, GenericObjective):
             raise RuntimeError("Bayesian optimiser requires a GenericObjective")
@@ -186,6 +192,12 @@ class BayesianBoTorch(Optimiser):
         self.ref_point = None if reference_point is None else -torch.tensor(reference_point)
         self.nadir_scale = nadir_scale
         self.pca_variance = pca_variance
+        self.num_restarts = num_restarts
+        self.raw_samples = raw_samples
+        self.mc_samples = mc_samples
+        self.batch_size = batch_size
+        self.sequential = bool(sequential)
+        self.num_fantasies = num_fantasies
         if acquisition is None:
             self.acquisition = default_acquisition(
                 objective.composition,
@@ -238,7 +250,6 @@ class BayesianBoTorch(Optimiser):
 
     def _get_candidates(
             self,
-            n_dim,
             bounds: np.ndarray,
             dataset: BayesDataset,
             test_dataset: BayesDataset,
@@ -259,19 +270,20 @@ class BayesianBoTorch(Optimiser):
                 cv_error = (posterior.mean - std_test_values).square().mean().item()
 
         # Build the acquisition function
-        acq, num_restarts, raw_samples = self._acq_func(dataset, model, n_dim)
+        acq = self._acq_func(dataset, model)
 
         # Optimise acquisition to find next candidate(s)
         candidates, _ = optimize_acqf(
             acq,
             bounds=torch.from_numpy(bounds.T).to(self.device).to(dataset.dtype),
             q=self.q,
-            num_restarts=num_restarts,
-            raw_samples=raw_samples,
+            num_restarts=self.num_restarts,
+            raw_samples=self.raw_samples,
+            sequential=self.sequential,
             options={
                 "sample_around_best": True,
                 "seed": self.seed,
-                "init_batch_limit": 4 * num_restarts,
+                "init_batch_limit": self.batch_size,
             },
         )
 
@@ -284,7 +296,7 @@ class BayesianBoTorch(Optimiser):
         # Multi-candidate: run cases in parallel
         with Pool(self.q) as pool:
             results = pool.map(lambda x: self.objective(x, concurrent=True), candidates)
-        return results
+        return list(results)
 
     def _get_random_points(
             self,
@@ -359,12 +371,9 @@ class BayesianBoTorch(Optimiser):
         self,
         dataset: BayesDataset,
         model: Model,
-        n_dim: int,
-    ) -> Tuple[AcquisitionFunction, int, int]:
+    ) -> AcquisitionFunction:
         # Default values for multi-restart optimisation
-        num_restarts = 12
-        raw_samples = max(256, 16 * n_dim * n_dim)
-        sampler = SobolQMCNormalSampler(torch.Size([512]), seed=self.seed)
+        sampler = SobolQMCNormalSampler(torch.Size([self.mc_samples]), seed=self.seed)
 
         # Find best value for the acquisition
         best = torch.max(self._composition(dataset.values, dataset.params)).item()
[…]
