### 2024-07-17 #6652 @hugoledoux
Now it was my turn to take too long to respond here (because of holidays).

I'm glad to see that both reviewers are enthusiastically recommending acceptance, and that you have improved your software/docs during the process. This is the idea of JOSS.

There is one issue that I would like to raise: the 4th author of the paper "FM Andrade Pires" doesn't have a single commit to the repository and is the corresponding author. JOSS is not like typical academic journals and have stricter rules for authorship, see here: https://joss.readthedocs.io/en/latest/submitting.html#authorship

@ruicoelhopedro Could you please let me know here how the 4th author was involved in the project? And also, I do believe that you should be the corresponding author for the paper, since it's about software and you're clearly the main person behind that software package.

And in any case, I think that
