### 2024-07-17 52b9737
Update references.bib
