+++ docs/joss_paper/references.bib
@@ -1,11 +1,11 @@
 
 @misc{nevergrad,
     author = {J. Rapin and O. Teytaud},
-    title = {{Nevergrad - A gradient-free optimization platform}},
+    title = {{Nevergrad --- A gradient-free optimization platform}},
     year = {2018},
     publisher = {GitHub},
     journal = {GitHub repository},
-    howpublished = {\url{https://GitHub.com/FacebookResearch/Nevergrad}}
+    howpublished = {\url{https://github.com/facebookresearch/nevergrad}}
 }
 
 
@@ -43,7 +43,7 @@ journal = {Journal of Open Source Software}
 @article{Coelho2023optm,
 title={Efficient constitutive parameter identification through optimisation-based techniques: A comparative analysis and novel composite Bayesian Optimisation Strategy},
 DOI={10.2139/ssrn.4674421},
-journal={submitted to Computational Methods in Applied Mathematics},
+journal={Computational Methods in Applied Mathematics},
 author={Cardoso Coelho, Rui and Carvalho Alves, A. Francisca and Andrade Pires, F. M.},
 year={2023}
 } 
