### 2024-07-17 #6652 @hugoledoux
I have proof-read the paper and I propose a few changes:
https://github.com/CM2S/piglot/pull/30
https://github.com/CM2S/piglot/pull/31

Also, the paper http://doi.org/10.2139/ssrn.4674421 was labelled as "submitted" but it seems accepted. Can you update the reference please?
