### 2024-07-17 #6652 @hugoledoux
@editoral check references
