### 2024-07-18 89a4a38
Further references update
