+++ docs/joss_paper/references.bib
@@ -22,7 +22,8 @@
     title={Optuna: A Next-generation Hyperparameter Optimization Framework},
     author={Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
     booktitle={Proceedings of the 25th {ACM} {SIGKDD} International Conference on Knowledge Discovery and Data Mining},
-    year={2019}
+    year={2019},
+    doi={10.48550/arXiv.1907.10902},
 }
 
 
@@ -40,12 +41,17 @@ journal = {Journal of Open Source Software}
 } 
 
 
-@article{Coelho2023optm,
-title={Efficient constitutive parameter identification through optimisation-based techniques: A comparative analysis and novel composite Bayesian Optimisation Strategy},
-DOI={10.2139/ssrn.4674421},
-journal={Computational Methods in Applied Mathematics},
-author={Cardoso Coelho, Rui and Carvalho Alves, A. Francisca and Andrade Pires, F. M.},
-year={2023}
+@article{Coelho2024optm,
+  title = {Efficient Constitutive Parameter Identification through Optimisation-Based Techniques: A Comparative Analysis and Novel Composite Bayesian Optimisation Strategy},
+  shorttitle = {Efficient Constitutive Parameter Identification through Optimisation-Based Techniques},
+  author = {Cardoso Coelho, R. P. and Carvalho Alves, A. Francisca and Andrade Pires, F. M.},
+  year = {2024},
+  journal = {Computer Methods in Applied Mechanics and Engineering},
+  volume = {427},
+  pages = {117039},
+  issn = {0045-7825},
+  doi = {10.1016/j.cma.2024.117039},
+  keywords = {Bayesian optimisation,Derivative-free optimisation,Inverse problems},
 } 
 
 
@@ -61,6 +67,7 @@ year={2023}
   primaryclass = {cs, math, stat},
   abstract = {Bayesian optimization provides sample-efficient global optimization for a broad range of applications, including automatic machine learning, engineering, physics, and experimental design. We introduce BoTorch, a modern programming framework for Bayesian optimization that combines Monte-Carlo (MC) acquisition functions, a novel sample average approximation optimization approach, auto-differentiation, and variance reduction techniques. BoTorch's modular design facilitates flexible specification and optimization of probabilistic models written in PyTorch, simplifying implementation of new acquisition functions. Our approach is backed by novel theoretical convergence results and made practical by a distinctive algorithmic foundation that leverages fast predictive distributions, hardware acceleration, and deterministic optimization. We also propose a novel "one-shot" formulation of the Knowledge Gradient, enabled by a combination of our theoretical and software contributions. In experiments, we demonstrate the improved sample efficiency of BoTorch relative to other popular libraries.},
   archiveprefix = {arXiv},
+  doi = {10.48550/arXiv.1910.06403},
   keywords = {Computer Science - Distributed; Parallel; and Cluster Computing,Computer Science - Machine Learning,Mathematics - Optimization and Control,Statistics - Machine Learning},
   file = {/home/rpcc/Zotero/storage/8VJMRIHM/Balandat et al_2020_BoTorch.pdf;/home/rpcc/Zotero/storage/AXQ2T5M5/1910.html}
 }
