### 2024-07-18 3468bbc
Update bad citation in paper.md
