### 2024-07-18 7d9f31f
a few typos fixed and commas added (#30)
