### 2024-07-18 63ddefd
JOSS review (#26)
Co-authored-by: FranciscaAlves <a.francisca.c.alves@gmail.com>
Co-authored-by: Hugo Ledoux <h.ledoux@tudelft.nl>
