### 2024-07-18 #6652 @ruicoelhopedro
> Now it was my turn to take too long to respond here (because of holidays).
>
> I'm glad to see that both reviewers are enthusiastically recommending acceptance, and that you have improved your software/docs during the process. This is the idea of JOSS.
>
> There is one issue that I would like to raise: the 4th author of the paper "FM Andrade Pires" doesn't have a single commit to the repository and is the corresponding author. JOSS is not like typical academic journals and have stricter rules for authorship, see here: https://joss.readthedocs.io/en/latest/submitting.html#authorship
>
> @ruicoelhopedro Could you please let me know here how the 4th author was involved in the project? And also, I do believe that you should be the corresponding author for the paper, since it's about software and you're clearly the main person behind that software package.
>
> And in any case, I think that

Thank you for the feedback!

Regarding the 4th author, even though they have not directly committed to the project, they have provided valuable input on the program structure and have been a key part of the development of piglot. Many design decisions were selected during brainstorming sessions based on their insight.

Also, the corresponding author has been updated, as suggested.

> I have proof-read the paper and I propose a few changes: [CM2S/piglot#30](https://github.com/CM2S/piglot/pull/30) [CM2S/piglot#31](https://github.com/CM2S/piglot/pull/31)
>
> Also, the paper http://doi.org/10.2139/ssrn.4674421 was labelled as "submitted" but it seems accepted. Can you update the reference please?

Thank you for the changes! I've addressed each one in its respective PR.
That paper has been accepted during this review process. The reference has been updated in the review PR, which has just been merged.
