+++ piglot/__init__.py
@@ -4,4 +4,4 @@ __title__ = 'piglot'
 __author__ = 'CM2S'
 __copyright__ = '2024, CM2S'
 __license__ = 'MIT'
-__version__ = '0.3.0'
+__version__ = '0.4.0'
