### 2024-07-22 #6652 @ruicoelhopedro
> @ruicoelhopedro
>
> At this point could you:
>
>     * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [x]  Please list the DOI of the archived version here.
>
>
> I can then move forward with recommending acceptance of the submission.

The version [v0.4.0](https://github.com/CM2S/piglot/releases/tag/v0.4.0) has been released. The software package has been published in https://doi.org/10.5281/zenodo.12796554
