### 2024-07-23 #6652 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
