### 2024-07-23 de96741
minor changes in casing in references.bib
