+++ docs/joss_paper/references.bib
@@ -36,13 +36,13 @@ publisher = {The Open Journal},
 volume = {8}, number = {87},
 pages = {5594},
 author = {Bernardo P. Ferreira and F. M. Andrade Pires and Miguel A. Bessa}, 
-title = {CRATE: A Python package to perform fast material simulations}, 
+title = {CRATE: A {P}ython package to perform fast material simulations}, 
 journal = {Journal of Open Source Software}
 } 
 
 
 @article{Coelho2024optm,
-  title = {Efficient Constitutive Parameter Identification through Optimisation-Based Techniques: A Comparative Analysis and Novel Composite Bayesian Optimisation Strategy},
+  title = {Efficient Constitutive Parameter Identification through Optimisation-Based Techniques: A Comparative Analysis and Novel Composite {B}ayesian Optimisation Strategy},
   shorttitle = {Efficient Constitutive Parameter Identification through Optimisation-Based Techniques},
   author = {Cardoso Coelho, R. P. and Carvalho Alves, A. Francisca and Andrade Pires, F. M.},
   year = {2024},
@@ -56,7 +56,7 @@ journal = {Journal of Open Source Software}
 
 
 @article{balandatBoTorchFrameworkEfficient2020,
-  title = {{{BoTorch}}: {{A Framework}} for {{Efficient Monte-Carlo Bayesian Optimization}}},
+  title = {{BoTorch}: A Framework for Efficient {M}onte-{C}arlo {B}ayesian Optimization},
   shorttitle = {{{BoTorch}}},
   author = {Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
   year = {2020},
@@ -74,7 +74,7 @@ journal = {Journal of Open Source Software}
 
 
 @article{cardosocoelhoMultiscaleModelCombining2023,
-  title = {A Multi-Scale Model Combining Martensitic Transformations with Multi-Phase Crystallographic Slip},
+  title = {A Multi-Scale Model Combining {M}artensitic Transformations with Multi-Phase Crystallographic Slip},
   author = {Cardoso Coelho, R. P. and {Vieira de Carvalho}, M. and Andrade Pires, F. M.},
   year = {2023},
   month = dec,
