### 2024-07-23 #6652 @danielskatz
@ruicoelhopedro - I'm suggesting some fairly minor casing changes in the references in https://github.com/CM2S/piglot/pull/33 - please merge this or let me know what you disagree with, then we can continue to acceptance and publication.
