### 2024-07-23 #6652 @ruicoelhopedro
> @ruicoelhopedro - I'm suggesting some fairly minor casing changes in the references in [CM2S/piglot#33](https://github.com/CM2S/piglot/pull/33) - please merge this or let me know what you disagree with, then we can continue to acceptance and publication.

Thank you, we agree and we have merged the suggested changes. Do we need to repeat the release and Zenodo dataset upload?
