### 2024-07-23 #6652 @danielskatz
No - JOSS doesn't require the paper source to be in the repo, since the paper itself (PDF and JATS) is archived separately.  I'll finish the process now.
