### 2024-07-23 #6652 @danielskatz
Congratulations to @ruicoelhopedro (Rui Pedro Cardoso Coelho) on your publication!!

JOSS depends on volunteers - we could't succeed without you
