### 2024-07-23 #6652 @ruicoelhopedro
Thank you all for the valuable feedback and input on our package! As our first time publishing in JOSS, we deeply appreciate the transparency, structured and quality-driven aspects of the revision process.
